Move branch_settings.php into admin dir
