+++ functions.php
@@ -23,4 +23,4 @@ require_once('core/load_more.php');
 require_once('core/main_news.php');
 
 require_once('core/infinite_scroll.php');
-require_once('branch_settings.php');
+require_once('admin/branch_settings.php');
